--- src/screens/General/Dashboard/index.js
@@ -1,3 +1,0 @@
-export { default as FavoriteCommittees } from './FavoriteCommittees';
-export { Leaderboard } from './Leaderboard';
-export { default as Dashboard } from './Dashboard';
